requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
requests-oauthlib>=1.3.0

# Optional but recommended
# For better SSL handling
certifi>=2023.0.0
